Prepares GitKit 2nd ed for publication. (#2)
**Pull Request Description**

Takes care of a number of last minute details before the 2nd ed is ready
for publication.

Closes #161
Closes #162
Closed #163


---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.