Prepares for publication of 2ed (#3)
**Pull Request Description**

Does a few things to prep for publication. Some of this was supposed to
be in pull #2 but didn't get pushed by accident.

Closes #162
Closes #163
Closes #194


---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.