Updates to the GitKit Text from January 2025 Sprint (#1)
**Pull Request Description**

All changes made in the January 2025 sprint will be merged to this
branch. Then this branch will be merged into `main` and a new text will
be created on Runestone from this Repo.

The most significant changes include:
- Migration to the [Codespaces
KitClient](https://github.com/HFOSSedu/KitClient-Codespace)
- Updates to lots of questions to make them autogradable.
- Change from using a merge tool, to editing raw merge info in VSCode.

Be sure to complete the following before merging:
- [x] - #117 
- [x] - #137
- [x] - #156
- [x] - #157
- [x] - #158 
- [x] - #161
- [x] - #162
- [x] - #163 
- [x] - #168 
- [x] - #194
---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

---------

Co-authored-by: lpostner <lori.postner@ncc.edu>
Co-authored-by: Karl R. Wurst <karl@w-sts.com>