Chapter 4 corrections (#37)
**Pull Request Description**

fixed typo in 4.3.2 and checked 4.3.3.1 and it was fine

Closes #35

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.