update Section 1.1.9 Deployting the GitKit (#49)
**Pull Request Description**

Updates instructions and images in "Section 1.1.9 Deploying the GitKit"
to be consistent with changes in the deployer.

Closes #23 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.