Updating Chapter 2 to add the text attribute to xrefs for sections (#50)
**Pull Request Description**

Added text type attributes for all xrefs to sections and subsections for
consistency.

Closes #16

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.