patches line 9 quotes error (#52)
**Pull Request Description**

Fixes a misplaced quote, and issues with the `text` attribute on some
`xrefs` that pointed to the "I would like to work on this!" text rather
than a numbered section and had custom text for the `xref`.

This allows the project to build without error. However, the CodeChat
preview still does not seem to be working. Will debug that separately.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.