Make image descriptions consistent (#46)
**Pull Request Description**

Make image descriptions consistent for accessibility. Short descriptions
are used for alt test and should be 100 characters or less. Descriptions
are only needed if more detail is required.

Closes #13

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.