Multiple choice question updates sections 5.1-5.5 (#54)
**Pull Request Description**

Updated multiple choice questions to include feedback for each choice
and hints.

Work associated with #17 but does not complete it.

Closes #17 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.