Mcupstreamingchanges (#57)
**Update all MC questions**

_MC questions all have feedback; Choose all questions have no feedback._

_Completes some of the work for #17 ._

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.