Ch. 2 additional updates (#56)
**Pull Request Description**
This pull requests formatted the pre-text files. It adds randomize
attributes to any choice one or choose all questions that were missing
them. Related to #17 but does not complete it.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.