Mcstayingsynch (#58)
**Updating MC Questions in Ch4**

_Ensured MC questions had feedback and Choose All questions had no
feedback_

_Updates to #17 ._

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

---------

Co-authored-by: braughtg <braught@dickinson.edu>